+++ parley/src/main/java/nu/parley/android/data/repository/DeviceRepository.java
@@ -1,5 +1,9 @@
 package nu.parley.android.data.repository;
 
+import android.content.Context;
+
+import java.util.UUID;
+
 import nu.parley.android.Parley;
 import nu.parley.android.data.model.Device;
 import nu.parley.android.data.net.Connectivity;
@@ -35,5 +39,17 @@ public final class DeviceRepository {
                 callback.onFailed(null, t.getMessage());
             }
         });
+    }
+
+    public static String getDeviceId(Context context) {
+        PreferenceRepository preferences = new PreferenceRepository();
+        String deviceId = preferences.getDeviceId(context);
+        if (deviceId == null) {
+            String newDeviceId = UUID.randomUUID().toString();
+            preferences.setDeviceId(context, newDeviceId);
+            return newDeviceId;
+        } else {
+            return deviceId;
+        }
     }
 }