+++ CHANGELOG.md
@@ -10,8 +10,8 @@ Parley now uses a unique device id per app installation as default setting.
 
 _What's changed?_
 
-- In Parley 3.4.5 and lower the Android DEVICE_ID was used as device id. This device id does not change per app installation, causing Anonymous chats to continue with their existing chat even when the user deleted the app and reinstalled the app. 
-- In Parley 3.5.0 and higher a random UUID is used as device id. This value is stored in the shared preferences by default and is generated once per installation. Updating the app will not result in a new device id as long as the cache isn't cleared. This ensures that anonymous chats always start empty.
+- In Parley 3.4.5 and lower the Android DEVICE_ID was used as device id. This device id does not change per app installation, causing anonymous chats to continue with their existing chat even when the user deleted the app and reinstalled the app. 
+- In Parley 3.5.0 and higher a random UUID is used as device id. This value is stored in the shared preferences by default and is generated once per installation. Updating the app won't result in a new device id as long as the cache isn't cleared. This ensures that anonymous chats always start empty.
 
 **Note**: This only affects the behavior of starting anonymous chats, chats that use the user authorization won't be affected by this change.